requirements.txt
google-api-python-client>=2.133.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
beautifulsoup4>=4.12.3
html5lib>=1.1
python-dotenv>=1.0.1
playwright>=1.47.0
pystray>=0.19.5
Pillow>=10.4.0
